// src/shared/api/model/action.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Gift App
 * Crypto Gift App contest
 * OpenAPI spec version: 1.0
 */
import type { ActionActionType } from './actionActionType';
import type { Gift } from './gift';
import type { User } from './user';

export interface Action {
  actionType: ActionActionType;
  gift?: Gift;
  recipient?: User;
  timestamp: string;
  user: User;
}