// src/shared/api/model/actionActionType.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Gift App
 * Crypto Gift App contest
 * OpenAPI spec version: 1.0
 */

export type ActionActionType = typeof ActionActionType[keyof typeof ActionActionType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ActionActionType = {
  send: 'send',
  buy: 'buy',
  receive: 'receive',
} as const;